+++ src/test/typescript/12-transpile.processing-instruction.test.js
@@ -7,10 +7,7 @@ describe('html-encoder-typescript: transpile.process-instruction', () => {
   }, true),
     '\n' +
     '    \n' +
-    '    node.appendChild((docElm => {\n' +
-    "    const node = docElm.appendChild(docElm.createProcessingInstruction('unknown',''));\n" +
-    '    return node;\n' +
-    '  })(self.docElm));;\n' +
+    "    document.createProcessingInstruction('unknown','');;\n" +
     '    \n' +
     '  '
   ));
@@ -27,9 +24,9 @@ describe('html-encoder-typescript: transpile.process-instruction', () => {
     '\n' +
     '    \n' +
     '    { \n' +
-    '          const fn = self.funcs.undefined.bind({},self, self.docElm, node);\n' +
+    '          const fn = funcs.undefined.bind({}, node);\n' +
     '\t\t\t\t\tconst startAt = node.childNodes.length;\n' +
-    "          const items = clone(self._getValue(self.data, 'undefined')) || [];\n" +
+    "          const items = clone(getValue(data, 'undefined')) || [];\n" +
     '\t\t\t\t\tconst nodes = fn(items);\n' +
     '\t\t\t\t\t\n' +
     '\t\t\t\t};\n' +
@@ -46,6 +43,6 @@ describe('html-encoder-typescript: transpile.process-instruction', () => {
   it('handles css instruction', () => assert.strictEqual(getProcessingInstruction({
     tag: 'css'
   }, true),
-    '\n    node.appendChild(self.docElm.createComment(`PI:css `));\n    ;\n    \n  '
+    '\n    node.appendChild(document.createComment(`PI:css `));\n    ;\n    \n  '
   ));
 });