+++ src/test/typescript/14-extract-functions.test.js
@@ -4,7 +4,7 @@ const { extractFunctions } = require('../../../dist/typescript/extract-functions
 describe('html-encoder-typescript: extract-functions', () => {
   it('handle empty if', () =>
     assert.strictEqual(extractFunctions({ type: 'ProcessingInstruction', tag: 'if' }, false, false),
-      'undefined (self, docElm, node) {\n' +
+      'undefined (node) {\n' +
       '          const fn = function () {  };\n' +
       '\t        return getAddedChildren(node, fn);\n' +
       '        }'
@@ -13,7 +13,7 @@ describe('html-encoder-typescript: extract-functions', () => {
 
   it('handle if condition', () =>
     assert.strictEqual(extractFunctions({ type: 'ProcessingInstruction', tag: 'if', attributes: { functionName: 'foo' } }, false, false),
-      'foo (self, docElm, node) {\n' +
+      'foo (node) {\n' +
       '          const fn = function () {  };\n' +
       '\t        return getAddedChildren(node, fn);\n' +
       '        }'
@@ -22,7 +22,7 @@ describe('html-encoder-typescript: extract-functions', () => {
 
   it('handle if condition ts ssr', () =>
     assert.strictEqual(extractFunctions({ type: 'ProcessingInstruction', tag: 'if', attributes: { functionName: 'foo' } }, true, true),
-      'foo (self:JSNode, docElm:Document, node:Node) {\n' +
+      'foo (node:Node) {\n' +
       '          const fn = function () {  };\n' +
       '\t        return getAddedChildren(node, fn);\n' +
       '        }'
@@ -31,13 +31,13 @@ describe('html-encoder-typescript: extract-functions', () => {
 
   it('handle empty foreach', () =>
     assert.strictEqual(extractFunctions({ type: 'ProcessingInstruction', tag: 'foreach' }, false, false),
-      'undefined (self, docElm, node, items) {\n' +
+      'undefined (node, items) {\n' +
       '          const fn = function() {\n' +
       '            \n' +
       '            \n' +
       '          };\n' +
       '\n' +
-      "          return self._forEach('undefined', 'undefined', node, fn, items);\n" +
+      "          return iterate(data, 'undefined', 'undefined', node, fn, items);\n" +
       '        }'
     )
   );