+++ src/typescript/index.ts
@@ -6,26 +6,21 @@ import { analyze } from './transpile.analyze';
 
 const encoding = 'utf-8';
 
-export type TargetType = 'js' | 'es' | 'ts' | 'code';
+export type TargetType = 'js' | 'es' | 'ts' | 'code' | 'es-code';
+
 export function transpile(instructions: Instruction, type: TargetType, isSSR: boolean = false): string {
   const functions = extractFunctions(instructions, type === 'ts', isSSR);
   let parsedString = getNode(instructions, isSSR);
   const { revivable, attr, css, data } = analyze(instructions);
   if (revivable) {
-    parsedString += `;self._defineSet();`;
+    parsedString += `;attachSetToNode(set, node);`;
   }
 
   return treeShake(
     readFileSync(getTemplateFile(__dirname, type), { encoding })
-      .replace(select('revive', revivable), '')
-      .replace(select('NodeType', revivable || attr || css), '')
-      .replace('/*! main-code-goes-here */', `((node) => { self.node = ${parsedString}; })(self.docElm);`)
-      .replace(select('any\-dynamic', revivable), '')
-      .replace(select('browser\-dynamic', revivable && !isSSR), '')
-      .replace(select('server\-dynamic', revivable && isSSR), '')
-      .replace(select('nodejs', (revivable && isSSR) || type === 'code'), '')
-      .replace(select('data', data), '')
+      .replace('const node = document.createTextNode(\'main-code-goes-here\');', `const node = ${parsedString};`)
       .replace('/*!funcs go here*/', functions)
+      .replace(select('data', data), '')
       .replace(select('funcs', functions.length > 0), '')
       .replace('//# sourceMappingURL=js-node.js.map', '')
       .replace(/(\r?\n){2,}/gm, '\n')
@@ -42,6 +37,7 @@ function getTemplateFile(folderName: String, type: TargetType): string {
     case 'ts':
       return `${template}.template-ts`;
     case 'es':
+    case 'es-code':
       return `${template}.es.js`;
     default:
       return `${template}.js`;
@@ -49,24 +45,16 @@ function getTemplateFile(folderName: String, type: TargetType): string {
 }
 
 function treeShake(code: string) {
-  return findFeatures(code)
-    .reduce(
-      (code: string, feature: string) => code.replace(select(`shakeable ${feature}`, isFeatureUsed(code, feature)), ''),
-      code
-    );
-}
-
-function isFeatureUsed(code: string, feature: string): boolean {
-  return (code.match(new RegExp(`${feature} = function|${feature}\\(|${feature}.bind`, 'gm')) || []).length > 1;
-}
-
-function findFeatures(code: string): string[] {
-  const featureFinder: RegExp = /\/\*!shakeable (\w*){\*\/\n/g;
-  const features: string[] = [];
-  let match: RegExpExecArray | null;
-  while ((match = featureFinder.exec(code)) !== null) {
-    features.push(match[1]);
+  const matches = /\/\*\!shakeable \{\*\/((.)*?)\/\*\!\} shakeable\*\//mg.exec(code);
+  if (matches === null) {
+    return code;
   }
 
-  return features;
+  const allFeatures = matches[1] ?? '';
+  const usedFeatures = allFeatures
+    .replace(/\s/g, '')
+    .split(',')
+    .filter(feature => (code.match(new RegExp(feature, 'g')) ?? []).length > 1)
+    .join(', ');
+  return code.replace(matches[0], usedFeatures);
 }