+++ src/typescript/transpile.nodes.ts
@@ -30,38 +30,38 @@ export function getDocument(instruction: Instruction, isSSR: boolean) {
   const name: string = instruction.attributes?.name as string;
   const publicId: string = instruction.attributes?.publicId as string || '';
   const systemId: string = instruction.attributes?.systemId as string || '';
-  return `(docElm => {
-    const node = docElm;
-    docElm.insertBefore(docElm.implementation.createDocumentType('${name}', '${publicId}', '${systemId}'), docElm.childNodes[0]);
-    docElm.removeChild(docElm.childNodes[1]);
+  return `(() => {
+    document.insertBefore(document.implementation.createDocumentType('${name}', '${publicId}', '${systemId}'), document.childNodes[0]);
+    document.removeChild(document.childNodes[1]);
+    const node = document;
     ${appendChildren(instruction.children, isSSR)}
-    return docElm;
-  })(self.docElm)`;
+    return document;
+  })()`;
 }
 
 export function getDocumentFragment(instruction: Instruction, isSSR: boolean) {
-  return `(docElm => {
-    const node = docElm.createDocumentFragment();
+  return `(() => {
+    const node = document.createDocumentFragment();
     ${instruction ? appendChildren(instruction.children, isSSR) : ''}return node;
-  })(self.docElm)`;
+  })()`;
 }
 
 export function getComment(instruction: Instruction) {
-  return `self.docElm.createComment(\`${instruction.value}\`)`;
+  return `document.createComment(\`${instruction.value}\`)`;
 }
 
 function getTextNode(instruction: Instruction) {
-  return `self.docElm.createTextNode(\`${instruction.value}\`)`;
+  return `document.createTextNode(\`${instruction.value}\`)`;
 }
 
 function getHTMLElement(instruction: Instruction, isSSR: boolean) {
   const attributes = (instruction.attributes) ? getAttributes(instruction.attributes) : '';
 
-  return `(docElm =>{
-    const node = docElm.createElement('${instruction.tag}');
+  return `(() =>{
+    const node = document.createElement('${instruction.tag}');
     ${attributes}${appendChildren(instruction.children, isSSR)}
     return node;
-  })(self.docElm)`;
+  })()`;
 }
 
 function appendChildren(children: Instruction[] = [], isSSR: boolean): string {